third_party/WebKit/LayoutTests/http/tests/encoding/meta-switch-mid-parse.html

<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
This test demonstrates that we don't buffer HTML documents looking for a meta
charset directive. Instead, we follow the HTML5 specification and switch
encoding mid-parse. The iframe below should alert two different encodings.
<iframe src="resources/pause-then-switch-encoding.cgi"></iframe>
